html/browsers/browsing-the-web/back-forward-cache/test2.html

<!doctype html>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helper.js"></script>
<script>
const localStoragePrefix = 'BFCache.test.';

const t = async_test();

clearData();
window.open('resources/events.sub.html', '_blank', 'noopener');

setInterval(t.step_func(() => {
    const result = getData('result');
    if (result) {
      assert_array_equals(JSON.parse(result), [
          'window.pagehide.persisted',
          'document.visibilitychange',
          'window.visibilitychange',
          // document.freeze/resume are not fired on Firefox.
          // 'document.freeze',
          // 'document.resume',
          'document.visibilitychange',
          'window.visibilitychange',
          'window.pageshow.persisted']);
      t.done();
    }
  }),
  100);
</script>
